pycook/cook.py: Add --pipe arg
Suppose Cookbook.py defines a (val f1 = fn : str -> str).

And the recipe `r1' is basically a shell command "foo-cmd --flags",
plus the output of it piped to `f1' and back to output.

Previously, the recipe would be shown as only:

    $ foo-cmd --flags

But that's not the whole of what the recipe did. We can now express it
more closely:

    $ foo-cmd --flags | cook --pipe f1