cook.el (cook-bury-buffer): Don't bury all the way down
Just take the top buffer from the full buffer list and replace the
current cook buffer.

I typically have 5-10 cook buffers around, with completion for
selecting them. Some of them are fresh and often used, others are
inactive for a long while. I'd like to keep the fresh ones closer to
the top, but that's not what the default `bury-buffer' does.