cook.el (cook): Detect :user_input
Due to some flaw in comint, we have to use "setsid -w" to get accurate
stdout in case there's a lot of it.

Then it works fine, except we lose stdin. Now a recipe can declare
that it will require user input by having an argument called
"user_input". In this case, we don't add "setsid -w".