Allow to edit the cmd in the comint-mode buffer and re-run it
* cook.el (cook-editing-command-p): Located on line 4 is the echo of the command that we just ran.
When dealing with complex recipes that take multiple args, it's convenient to be able to just edit
one arg and re-run instead of re-entering all those args from the start.

(cook-arg-history): The args are now entered in the minibuffer when called from Elisp.  Add a
history var to make entering args more convenient.